+++ src/main/java/com/feelory/feelory_backend/global/security/auth/dto/response/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.feelory.feelory_backend.global.security.auth;
+package com.feelory.feelory_backend.global.security.auth.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
